requirements.txt
#Python 3.12.3

asgiref==3.10.0
Django==5.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
PyJWT==2.10.1
sqlparse==0.5.3
